+++ extension/src/views/gradleDaemons/GradleDaemonTreeItem.ts
@@ -1,8 +1,8 @@
 import * as vscode from "vscode";
 import * as path from "path";
-//import { DaemonInfo } from "../../proto/gradle_pb";
 import { DAEMON_ICON_MAP } from "../constants";
-import { DaemonInfo, DaemonStatus } from "./models/DaemonInfo";
+import { DaemonInfo } from "./models/DaemonInfo";
+import { DaemonStatus } from "./models/DaemonStatus";
 interface StatusEnumMapByValue {
     [key: number]: string;
 }
