+++ extension/src/views/gradleDaemons/models/DaemonStatus.ts
@@ -1,0 +1,7 @@
+export enum DaemonStatus {
+    IDLE = 0,
+    BUSY = 1,
+    STOPPED = 2,
+    STOPPING = 3,
+    CANCELED = 4,
+}
