+++ extension/src/views/gradleDaemons/models/GradleConnectionType.ts
@@ -1,0 +1,5 @@
+export enum GradleConnectionType {
+    WRAPPER,
+    LOCALINSTALLATION,
+    SPECIFICVERSION
+}
